requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.104.1
h11==0.16.0
idna==3.11
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.24.0
beautifulsoup4
